Small refactoring, added debug adapter
