+++ src/Draco.DebugAdapter/Program.cs
@@ -25,6 +25,8 @@ internal static class Program
 {
     internal static async Task Main(string[] args)
     {
+        throw new InvalidOperationException();
+
         var stdioFlag = new Option<bool>(name: "--stdio", description: "A flag to set the transportation option to stdio");
 
         var runCommand = new Command("run", "Runs the debug adapter")
